Make reboot support a bit more backward compatible
Both beaker [1] and restraint [2] documentation mention the
`REBOOTCOUNT` as the variable used to store number of reboots.
Keep the backward-compatible naming consistent to make the test
migration as simple and straightforward as possible.

Add support for `TMT_REBOOT_COUNT` and `RSTRNT_REBOOTCOUNT`
variables as well and document them. Slighly modify the reboot
implementation to always define the reboot variables (as it is
done in restraint/beaker). Adjust tests and docs accordingly.

[1] https://beaker-project.org/docs/user-guide/task-environment.html#envvar-REBOOTCOUNT
[2] https://restraint.readthedocs.io/en/latest/remove_rhts.html#legacy-rhts-task-environment-variables

Resolves #991.